LayoutTests/imported/w3c/web-platform-tests/test-scenarios/17.1-crashing-test-now-passes-expected.txt

CONSOLE MESSAGE: line 13: ReferenceError: Can't find variable: cloningTest

FAIL IndexedDB: large nested objects are cloned correctly ReferenceError: Can't find variable: cloningTest
